Add detail to CLI directions
Listing out the steps needed to install HACS will allow users to install 
HACS
on their own.

Example commands are given, but the exact steps will depend on the 
specifics
of the user's setup.

If a user is interested in doing these steps manually, the user will 
most likely
know how their Home Assistant is set up and also how to change the 
example
commands for their own environment.