Add step to set correct permissions
The user account running Home Assistant needs to be able to write into the
custom_components directory.

If the custom_components directory is created by a different user account, or if
the hacs.zip file is extracted by a different user, then the resulting
directories may not be writable by the Home Assistant user account.

This step ensures that the directories are owned by the Home Assistant user
account, which in nearly all situations, would allow writing.